requirements.txt
astroid==2.0.4
Django==2.1.3
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pipfile==0.0.2
psycopg2==2.7.5
pylint==2.1.1
pytz==2018.7
six==1.11.0
toml==0.10.0
typed-ast==1.1.0
wrapt==1.10.11
